+++ ts/examples/chat/src/common.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { openai } from "aiclient";
+import { ChatModel, openai, TextEmbeddingModel } from "aiclient";
 import { ChalkInstance } from "chalk";
 import {
     ArgDef,
@@ -27,6 +27,46 @@ import {
 import { KnowledgeProcessorWriter } from "./knowledgeProc/knowledgeProcessorWriter.js";
 import path from "path";
 import fs from "fs";
+import * as knowLib from "knowledge-processor";
+
+/**
+ * Models used by example code
+ */
+export type Models = {
+    chatModel: ChatModel;
+    answerModel: ChatModel;
+    embeddingModel: TextEmbeddingModel;
+    embeddingModelSmall?: TextEmbeddingModel | undefined;
+};
+
+export function createModels(): Models {
+    const chatModelSettings = openai.apiSettingsFromEnv(openai.ModelType.Chat);
+    chatModelSettings.retryPauseMs = 10000;
+    const embeddingModelSettings = openai.apiSettingsFromEnv(
+        openai.ModelType.Embedding,
+    );
+    embeddingModelSettings.retryPauseMs = 25 * 1000;
+
+    const models: Models = {
+        chatModel: openai.createJsonChatModel(chatModelSettings, [
+            "chatMemory",
+        ]),
+        answerModel: openai.createChatModel(),
+        embeddingModel: knowLib.createEmbeddingCache(
+            openai.createEmbeddingModel(embeddingModelSettings),
+            1024,
+        ),
+        /*
+        embeddingModelSmall: knowLib.createEmbeddingCache(
+            openai.createEmbeddingModel("3_SMALL", 1536),
+            256,
+        ),
+        */
+    };
+    models.chatModel.completionSettings.seed = 123;
+    models.answerModel.completionSettings.seed = 123;
+    return models;
+}
 
 export async function pause(ms: number): Promise<void> {
     return new Promise((resolve) => setTimeout(resolve, ms));
